Equipment.h and Equipment.cpp created
digging and money multipliers, processed mine, get, add, spend and initial values ​​were created, some bugs in miner.cpp and miner.h were fixed

+++ include/Miner.h
@@ -6,7 +6,7 @@
 class Miner {
 private:
     // Player's values
-    int money_;
+    float money_;
 
     int health_;
     int energy_;
